README: Fix formatting of Overview header
This patch adds a newline between the VisionCpp header logo and the Overview header text, in order to fix parsing of the README as GitHub Flavored Markdown. As it stands, the Overview header text is currently not parsed as a header, and the `#` symbol is left untouched by the parser. Additionally, the "VisionCpp" text on the following line is not parsed as bold, and the asterisks are left untouched.